Rename core-inlineblocks directory to core-inline-blocks
Previously I had used only a single dash, in preparation for adding to
webpack where only the first dash was being replaced by the
`camelCaseDash` function. However, now that this function has been
updated to replace all dashes, this is no longer a limitation.
