+++ core-inline-blocks/inline-image.js
